Fix spurious test failure in test_obtain_credentials_from_stdin, and move make_credentials to a shared location.
